+++ Data/Align/Key.hs
@@ -1,5 +1,5 @@
 -----------------------------------------------------------------------------
--- | Module     :  Data.Aligned.Key
+-- | Module     :  Data.Align.Key
 --
 -- 'These'-based zipping and unzipping of indexed functors.
 --
@@ -15,7 +15,7 @@ import Data.Align
 import Data.These
 
 -- Instances
---import Control.Applicative  (ZipList)
+import Control.Applicative  (ZipList)
 import Data.Hashable        (Hashable)
 import Data.HashMap.Strict  (HashMap)
 import Data.IntMap          (IntMap)
@@ -33,7 +33,9 @@ class (Keyed f, Align f) => AlignWithKey f where
 
 instance AlignWithKey Maybe
 instance AlignWithKey []
---instance AlignWithKey ZipList
+
+-- | @since 0.7.6
+instance AlignWithKey ZipList
 instance AlignWithKey Seq
 instance AlignWithKey IntMap
 instance Ord k => AlignWithKey (Map k)